requirements.txt
aiogram
requests
psycopg2
python-binance
pandas
numpy
ccxt

